% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitpal.R
\name{bitpal}
\alias{bitpal}
\title{Get a fun color palette}
\usage{
bitpal(pal, n)
}
\arguments{
\item{pal}{Name of the color palette ("dunkin", "seinfeld", or "monopoly")}

\item{n}{Number of colors to return}
}
\value{
A character vector of hex codes
}
\description{
Color palettes based on Dunkin Donuts, Seinfeld, and Monopoly.
}
\examples{
# Retrieve some bitpals
bitpal(pal="monopoly", n=7)
bitpal(pal="dunkin", n=4)
bitpal(pal="seinfeld", n=5)
# Use some bitpals
data <- data.frame(product=c("Strawberry spinkle donut", "Iced coffee", "Munchkin", "Boston creme"),
                   points=c(100, 85, 80, 10))
ggplot(data, aes(y=product, x=points, fill=product)) +
  geom_bar(stat="identity") +
  labs(x="Deliciousness points", y="", title="Tracey's opinion of Dunkin products") +
  scale_fill_manual(values=bitpal("dunkin", n=nrow(data)), guide=F) +
  theme_bw()
}
